\name{coerce-methods}
\docType{methods}
\alias{coerce-methods}
\alias{coerce,RegistryKeyPath,character-method}
\title{Coerce method for registry path identifier to a string}
\description{
 This method simply turns a \code{RegistryKeyPath}
 into its equivalent, expanded string form.
}
\section{Methods}{
\describe{
\item{from = "RegistryKeyPath", to = "character"}{ converts a path to a string with elements separated by \\\\ }
}}
\keyword{methods}
\keyword{utilities}

