\name{resolveKey}
\alias{resolveKey}
\title{Verify key exists in registry}
\description{
 This function and its methods provide a way to organize a key specification into a
\code{RegistryKeyPath} and verify that it exists within the registry, identifying a valid key entry.
}
\usage{
resolveKey(path = "", top = .BuiltinKeys[1])
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{path}{the identifier for a registry key.}
  \item{top}{the top-level key in which the path specification 
    is found.}
}
\value{

}

\references{
 \url{http://www.omegahat.org/SWinRegistry}
}
\author{Duncan Temple Lang <duncan@wald.ucdavis.edu>}


\examples{

}
\keyword{utilities}

