\name{getRegistryKeyValues-methods}
\docType{methods}
\alias{getRegistryKeyValues-methods}
\alias{getRegistryKeyValues,RegistryKeyPath,missing-method}
\alias{getRegistryKeyValues,character-method}
\alias{getRegistryValue-methods}
\alias{getRegistryValue,character-method}
\alias{getRegistryValue,RegistryKeyPath-method}
\title{Value of one or more keys in the Windows registry}
\description{
These methods provide different ways to specify the path
of interest to fetch the name-value pairs or
single value in the Windows registry.
}
\section{Methods}{
\describe{

\item{path = "RegistryKeyPath", top = "missing"}{ ~~describe this method here }
\item{path = "character"}{ ~~describe this method here }

% getRegistryValue
\item{path = "character"}{ ~~describe this method here }
\item{path = "RegistryKeyPath"}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{utilities}

